+++ src/map/map_buffer.c
@@ -15,8 +15,8 @@
 static void				copy_line(t_mapbuffer *node,
 		enum e_map_tile_type *mapdata)
 {
-	int		i;
-	char	c;
+	unsigned int	i;
+	char			c;
 
 	i = 0;
 	while (i < node->str_size)
@@ -31,7 +31,7 @@ static void				copy_line(t_mapbuffer *node,
 	}
 }
 
-t_mapbuffer				*mbuf_create(char *line, int line_size)
+t_mapbuffer				*mbuf_create(char *line, unsigned int line_size)
 {
 	t_mapbuffer	*buff;
 
@@ -57,14 +57,15 @@ void					mbuf_append(t_mapbuffer *root, t_mapbuffer *new)
 
 enum e_map_tile_type	*mbuf_finalize(t_mapbuffer *root, int *map_height)
 {
-	int						line_size;
-	int						lines;
-	int						i;
+	unsigned int			line_size;
+	unsigned int			lines;
+	unsigned int			i;
 	enum e_map_tile_type	*mapdata;
 
 	line_size = root->line_size_max;
 	lines = root->lines;
-	mapdata = ft_calloc(lines, line_size * sizeof(enum e_map_tile_type));
+	mapdata = ft_calloc((size_t)lines, line_size *
+		sizeof(enum e_map_tile_type));
 	if (!mapdata)
 		return (NULL);
 	i = 0;
